Separated ProgressWithIndicator from original Progress
