+++ src/components/Progress/Progress.js
@@ -1,33 +1,16 @@
 import React, { useState } from 'react'
 import propTypes from 'prop-types'
 import classNames from 'classnames'
-import { ResultIndicator } from '../ResultIndicator'
-import languageService from '../../services/language'
 import styles from './Progress.module.scss'
-
-const getTranslation = word => languageService.getTranslation(word)
 
 const Progress = ({
   value,
   variant,
   indeterminate,
   className,
-  withText,
-  tiny,
-  success,
-  error,
-  errorMessage,
-  successMessage,
-  inQueue,
   ...otherProps
 }) => {
   const [circleRef, setCircleRef] = useState()
-
-  const progressText = withText && !indeterminate && (
-    <h5 className={styles.progressText}>
-      {!tiny && getTranslation('processing')} {value}%
-    </h5>
-  )
 
   const renderProgressCircle = () => {
     const classes = classNames(
@@ -87,24 +70,7 @@ const Progress = ({
     )
   }
 
-  const feedbackProps = {
-    tiny,
-    error,
-    success,
-    errorMessage,
-    successMessage,
-    inQueue,
-  }
-  const feedbackOptions = [error, success, inQueue].filter(Boolean)
-
-  return feedbackOptions.length ? (
-    <ResultIndicator {...feedbackProps} />
-  ) : (
-    <>
-      {progressText}
-      {variant === 'linear' ? renderProgressLine() : renderProgressCircle()}
-    </>
-  )
+  return variant === 'linear' ? renderProgressLine() : renderProgressCircle()
 }
 
 Progress.defaultProps = {
@@ -123,20 +89,6 @@ Progress.propTypes = {
   value: propTypes.number,
   /** Indeterminate mode. */
   indeterminate: propTypes.bool,
-  /** With text above progress. For use with min-width 150px, otherwise using tiny prop. */
-  withText: propTypes.bool,
-  /** Determines if text with `processing` or not. */
-  tiny: propTypes.bool,
-  /** Determines if the action was done successfully. */
-  success: propTypes.bool,
-  /** Determines if the action fails. */
-  error: propTypes.bool,
-  /** Message to show when action fails.*/
-  errorMessage: propTypes.string,
-  /** Message to show when action succeeds.*/
-  successMessage: propTypes.string,
-  /** Action waits to be done.*/
-  inQueue: propTypes.bool,
 }
 
 export default Progress
