+++ src/components/ProgressWithIndicator/index.js
@@ -1,0 +1,1 @@
+export { default as ProgressWithIndicator } from './ProgressWithIndicator'
