+++ src/index.js
@@ -27,5 +27,6 @@ export { Progress } from './components/Progress'
 export { Tabs } from './components/Tabs'
 export { RangeSlider } from './components/RangeSlider'
 export { ResultIndicator } from './components/ResultIndicator'
+export { ProgressWithIndicator } from './components/ProgressWithIndicator'
 
 export { useFormProvider } from './components/FormProvider/useFormProvider'
